remove all code related to veto tracking
veto tracking was a special mode of tracking a beatmap, meant for BNs after they could veto a map. fortunately, the osu website's notification system has evolved quick enough for this feature to have no real use, and it probably never have a use, so this feature has been removed in favor of having less code, less to maintain.